pyproject.toml
[tool.poetry]
name = "qr-app"
version = "0.1.0"
description = ""
authors = ["gal16v8d"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
tk = "^0.1.0"
pyzbar = "^0.1.9"
qrcode = "^7.4.2"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
